extract component functionality of RobotInteractionCockpit and put it in seperate file
